requirements.txt
gradio==3.9.1
nmslib==2.1.1
numpy==1.19.5
pandas==1.1.4
plotly==5.11.0
scikit_learn==0.24.0
sentence_transformers==2.2.2
umap_learn==0.4.6
